Throw error for too many decimals
Add in another conditional into both the deposit and withdraw functions
so that when the user inputs an amount with more than two decimals,
an error is thrown instead of adding or subtracting fractions of a
penny from the balance.

Achieve this by checking whether the amount to two decimal places is
the same value as the amount itself using the toFixed function. The
balance is also checked in the tests to make sure the amount has not
been added or subtracted.