Throw an error for invalid dates
Add another conditional to the deposit and withdraw functions to
throw an error when an invalid day, month or year is given by the user.
Transactions with invalid dates should not go through and should not
be added to the transaction log.

In a private method, format the date into a string that can be passed
into new Date() as an argument. getTime() will return false for an
invalid date but will return true for a valid date. This can therefore
be used to check if the date string passed into deposit and withdraw is
a valid date.