Make minor aesthetic changes
Change the name of the instance variable that holds the transactions
in the TransactionLog class in order to avoid any confusion with the
log instance variable in the Account class.

Change the error thrown when an invalid amount is given in the deposit
and withdraw functions so that the user knows it was an error with
the amount entered.