Format transaction history in the statement
Create the Statement class to take the transaction log as an argument
in the formatLog function and use the data in the transaction objects
in its history to present the accounts statement in a readable way
for the user.

Just account for one deposit for now. Add in functionality for
withdrawals and multiple transactions next.

Also, change the push method in the transactionLog class to unshift
so that the latest transactions are first in the array. This is so
that the latest transactions will be first on the statement when
iterating through the history array in the Statement class.