Create statement for multiple transactions
Modify the formatLog function to iterate through the history array
from the transaction log and add a new line to the statement for each
transaction. Add a conditional to distinguish between deposits and
withdrawals so that the amount for each type of transaction is under
the correct column in the statement.

Also, display the date each transaction in the statement occurred and
the updated balance.