Connect Statement class to Account class
Add a new instance varibale in the Account class and use dependency
injection to make this new variable the accounts statement.

Create the printStatement function which simply calls the statement
to format the transaction log and return the resulting statement.

Also, make some minor changes to the mock implementations to exactly
which objects are gaining the mocked function.