Fix printStatement and balance float
Modify the printStatement function so that the statement is printed
to the console. This was not being printed in the correct format before.

Also, fix a minor bug due to the way Javascript estimates floats. At
times, the balance can become a recurring decimal when particular
amounts are added or taken from it. Fix this by using toFixed(2) to
round the balance to 2 decimal places and then use parseFloat() to
convert it back into a float.