Add error when amount is negative or 0
Modify the amountCheck private method so another error is thrown when
the user attempts to deposit or withdraw an amount that is less than
or equal to 0.

Also, change the error message for a number not being passed for the
amount so that the user can clearly see what was wrong with their
input.