Add error for date format
Modify the dateCheck private method to throw another error when the
user does not enter a date in the format DD/MM/YYYY. Do this by checking
that the second and fifth index in the date string are both "/". This is
so that the dates are all consistently in the same format when printing
the statement.