Update README and displayBalance
Add the displayBalance into the How To Use section. Also, change the
output of the function so it prints the balance using console.log.
This is so that it follows the same consistency as the console
messages in the deposit, withdraw and printStatement functions which
the user will be runnning.