Update date to use Date object
Modify the statement and transaction log classes to incorporate actual
date objects as oppose to using a string. Therefore remove the date check
in the transaction log class as the date will be added automatically.
Use toLocaleDateString in order to format the date into a short format
on the statement.