Make multiple changes
In the account class tests, check that the correct arguments are being sent
to the transaction log and statement rather than just solely check the
balance. Then remove two unnecessary tests that specifically tested for
this since this will now be incorporated to each test.

In the statement class, remove the dependency of the transaction log class
by simply passing the getHistory method of the log into the formatLog
function of the statement class.

In the transaction log class, remove the balance argument passed into
addTransaction as this will now be calculated from the previous
transaction.