Remove this.balance
Remove this.balance from the account class and change displayBalance
to read the updated balance of the last transaction in order to
function. Update tests in account.test.js to no longer check the
balance displayed now and just check that addTransaction is being
called correctly.

Update #balanceCheck to no longer rely of a balance variable as well.